Move update_weights_from_distributed onto WeightUpdater
